+++ dataiter/test/test_list_of_dicts.py
@@ -336,20 +336,43 @@ class TestListOfDicts:
         assert len(data) == 35
         assert sum(data.pluck("downloads")) == 18226489
 
-    def test_sort(self):
+    def test_sort_ascending(self):
         orig = test.list_of_dicts("downloads.json")
         data = orig.sort(date=1, category=1)
         assert len(data) == len(orig)
         assert all(x in orig for x in data)
+        assert data.pluck("date") == sorted(data.pluck("date"))
+        assert data[0].date == min(data.pluck("date"))
+        assert data[-1].date == max(data.pluck("date"))
+        assert data[0].category == min(data.pluck("category"))
+        assert data[-1].category == max(data.pluck("category"))
 
-    def test_sort_with_none(self):
+    def test_sort_descending(self):
+        orig = test.list_of_dicts("downloads.json")
+        data = orig.sort(date=-1, category=-1)
+        assert len(data) == len(orig)
+        assert all(x in orig for x in data)
+        assert data.pluck("date") == sorted(data.pluck("date"), reverse=True)
+        assert data[0].date == max(data.pluck("date"))
+        assert data[-1].date == min(data.pluck("date"))
+        assert data[0].category == max(data.pluck("category"))
+        assert data[-1].category == min(data.pluck("category"))
+
+    def test_sort_with_none_ascending(self):
         # Nones should be sorted last.
         orig = test.list_of_dicts("downloads.json")
         orig[0].category = None
         data = orig.sort(category=1)
         assert data[-1] is orig[0]
 
-    def test_sort_with_none_multiple_keys(self):
+    def test_sort_with_none_descending(self):
+        # Nones should be sorted last.
+        orig = test.list_of_dicts("downloads.json")
+        orig[0].category = None
+        data = orig.sort(category=-1)
+        assert data[-1] is orig[0]
+
+    def test_sort_with_none_multiple_keys_ascending(self):
         # Nones should be sorted group-wise last.
         orig = test.list_of_dicts("downloads.json")
         orig[0].category = None
@@ -358,6 +381,18 @@ class TestListOfDicts:
         orig[2].date = None
         data = orig.sort(date=1, category=1)
         assert data[ 4] is orig[0]
+        assert data[-2] is orig[1]
+        assert data[-1] is orig[2]
+
+    def test_sort_with_none_multiple_keys_descending(self):
+        # Nones should be sorted group-wise last.
+        orig = test.list_of_dicts("downloads.json")
+        orig[0].category = None
+        orig[1].date = None
+        orig[2].category = None
+        orig[2].date = None
+        data = orig.sort(date=-1, category=-1)
+        assert data[-3] is orig[0]
         assert data[-2] is orig[1]
         assert data[-1] is orig[2]